🌐 convert to global function
