+++ src/classes/Commands/sub-classes/PricelistManager.ts
@@ -7,7 +7,7 @@ import Currencies from 'tf2-currencies-2';
 import pluralize from 'pluralize';
 import dayjs from 'dayjs';
 import sleepasync from 'sleep-async';
-import { removeLinkProtocol, testSKU, getItemFromParams } from '../functions/utils';
+import { removeLinkProtocol, testSKU, getItemFromParams, fixSKU } from '../functions/utils';
 import Bot from '../../Bot';
 import CommandParser from '../../CommandParser';
 import { Entry, EntryData, PricelistChangedSource } from '../../Pricelist';
@@ -158,7 +158,7 @@ export default class PricelistManagerCommands {
             params.sku = SKU.fromObject(item);
         }
 
-        params.sku = this.fixSKU(params.sku);
+        params.sku = fixSKU(params.sku);
 
         this.bot.pricelist
             .addPrice(params as EntryData, true, PricelistChangedSource.Command)
@@ -171,18 +171,6 @@ export default class PricelistManagerCommands {
             .catch(err => {
                 this.bot.sendMessage(steamID, `❌ Failed to add the item to the pricelist: ${(err as Error).message}`);
             });
-    }
-
-    private fixSKU(sku: string): string {
-        if (sku.includes(';15') && sku.includes(';strange')) {
-            // Only fix for Strange War Paint/Skins and Strange Unusual War Paint/Skins (weird variant)
-            const item = SKU.fromString(sku);
-            item.quality = 11;
-            item.quality2 = null;
-            return SKU.fromObject(item);
-        }
-
-        return sku;
     }
 
     private generateAddedReply(bot: Bot, isPremium: boolean, entry: Entry): string {
@@ -803,7 +791,7 @@ export default class PricelistManagerCommands {
             params.sku = SKU.fromObject(item);
         }
 
-        params.sku = this.fixSKU(params.sku);
+        params.sku = fixSKU(params.sku);
 
         if (!this.bot.pricelist.hasPrice(params.sku as string)) {
             return this.bot.sendMessage(steamID, '❌ Item is not in the pricelist.');
@@ -1124,7 +1112,7 @@ export default class PricelistManagerCommands {
             params.sku = SKU.fromObject(item);
         }
 
-        params.sku = this.fixSKU(params.sku);
+        params.sku = fixSKU(params.sku);
 
         this.bot.pricelist
             .removePrice(params.sku as string, true)
@@ -1183,7 +1171,7 @@ export default class PricelistManagerCommands {
             return this.bot.sendMessage(steamID, '❌ Missing item');
         }
 
-        params.sku = this.fixSKU(params.sku);
+        params.sku = fixSKU(params.sku);
 
         const match = this.bot.pricelist.getPrice(params.sku as string);
         if (match === null) {
